extract relationship out into its own file
